moved helper functions to separate files, cleaned up other tests
