+++ public/expressions/__tests__/overlay_anomalies.test.ts
@@ -4,18 +4,19 @@
  */
 
 import { setClient } from '../../services';
-import { httpClientMock, coreServicesMock } from '../../../test/mocks';
+import { httpClientMock } from '../../../test/mocks';
 import {
   convertAnomaliesToPointInTimeEventsVisLayer,
   getAnomalies,
   getVisLayerError,
   getDetectorResponse,
-} from '../overlay_anomalies';
+} from '../helpers';
 import {
-  anomalyResultSummary,
-  noAnomaliesResultResponse,
-  parsedAnomalies,
-  selectedDetectors,
+  ANOMALY_RESULT_SUMMARY,
+  ANOMALY_RESULT_SUMMARY_DETECTOR_ID,
+  NO_ANOMALIES_RESULT_RESPONSE,
+  PARSED_ANOMALIES,
+  SELECTED_DETECTORS,
 } from '../../pages/utils/__tests__/constants';
 import {
   DETECTOR_HAS_BEEN_DELETED,
@@ -31,27 +32,27 @@ describe('overlay_anomalies spec', () => {
 
   const ADPluginResource = {
     type: VIS_LAYER_PLUGIN_TYPE,
-    id: 'hNX8l4ABuV34PY9I1EAZ',
+    id: ANOMALY_RESULT_SUMMARY_DETECTOR_ID,
     name: 'test-1',
-    urlPath: `${PLUGIN_NAME}#/detectors/hNX8l4ABuV34PY9I1EAZ/results`, //details page for detector in AD plugin
+    urlPath: `${PLUGIN_NAME}#/detectors/${ANOMALY_RESULT_SUMMARY_DETECTOR_ID}/results`, //details page for detector in AD plugin
   };
 
   describe('getAnomalies()', () => {
     test('One anomaly', async () => {
-      httpClientMock.post = jest.fn().mockResolvedValue(anomalyResultSummary);
+      httpClientMock.post = jest.fn().mockResolvedValue(ANOMALY_RESULT_SUMMARY);
       const receivedAnomalies = await getAnomalies(
-        'hNX8l4ABuV34PY9I1EAZ',
+        ANOMALY_RESULT_SUMMARY_DETECTOR_ID,
         1589258564789,
         1589258684789
       );
-      expect(receivedAnomalies).toStrictEqual(parsedAnomalies);
+      expect(receivedAnomalies).toStrictEqual(PARSED_ANOMALIES);
     });
     test('No Anomalies', async () => {
       httpClientMock.post = jest
         .fn()
-        .mockResolvedValue(noAnomaliesResultResponse);
+        .mockResolvedValue(NO_ANOMALIES_RESULT_RESPONSE);
       const receivedAnomalies = await getAnomalies(
-        'hNX8l4ABuV34PY9I1EAZ',
+        ANOMALY_RESULT_SUMMARY_DETECTOR_ID,
         1589258564789,
         1589258684789
       );
@@ -60,7 +61,7 @@ describe('overlay_anomalies spec', () => {
     test('Failed response', async () => {
       httpClientMock.post = jest.fn().mockResolvedValue({ ok: false });
       const receivedAnomalies = await getAnomalies(
-        'hNX8l4ABuV34PY9I1EAZ',
+        ANOMALY_RESULT_SUMMARY_DETECTOR_ID,
         1589258564789,
         1589258684789
       );
@@ -71,21 +72,21 @@ describe('overlay_anomalies spec', () => {
     test('get detector', async () => {
       httpClientMock.get = jest
         .fn()
-        .mockResolvedValue({ ok: true, response: selectedDetectors[0] });
+        .mockResolvedValue({ ok: true, response: SELECTED_DETECTORS[0] });
       const receivedAnomalies = await getDetectorResponse(
         'gtU2l4ABuV34PY9ITTdm'
       );
       expect(receivedAnomalies).toStrictEqual({
         ok: true,
-        response: selectedDetectors[0],
+        response: SELECTED_DETECTORS[0],
       });
     });
   });
   describe('convertAnomaliesToPointInTimeEventsVisLayer()', () => {
     test('convert anomalies to PointInTimeEventsVisLayer', async () => {
       const expectedTimeStamp =
-        parsedAnomalies[0].startTime +
-        (parsedAnomalies[0].endTime - parsedAnomalies[0].startTime) / 2;
+        PARSED_ANOMALIES[0].startTime +
+        (PARSED_ANOMALIES[0].endTime - PARSED_ANOMALIES[0].startTime) / 2;
       const expectedPointInTimeEventsVisLayer = {
         events: [
           {
@@ -95,17 +96,16 @@ describe('overlay_anomalies spec', () => {
         ],
         originPlugin: 'anomalyDetectionDashboards',
         pluginResource: {
-          id: 'hNX8l4ABuV34PY9I1EAZ',
+          id: ANOMALY_RESULT_SUMMARY_DETECTOR_ID,
           name: 'test-1',
           type: 'Anomaly Detectors',
-          urlPath:
-            'anomaly-detection-dashboards#/detectors/hNX8l4ABuV34PY9I1EAZ/results',
+          urlPath: `anomaly-detection-dashboards#/detectors/${ANOMALY_RESULT_SUMMARY_DETECTOR_ID}/results`,
         },
         type: 'PointInTimeEvents',
       };
       const pointInTimeEventsVisLayer =
         await convertAnomaliesToPointInTimeEventsVisLayer(
-          parsedAnomalies,
+          PARSED_ANOMALIES,
           ADPluginResource
         );
       expect(pointInTimeEventsVisLayer).toStrictEqual(
@@ -123,7 +123,6 @@ describe('overlay_anomalies spec', () => {
         message: error.message,
       };
       const visLayerError = await getVisLayerError(error);
-      console.log('receivedError: ' + JSON.stringify(visLayerError));
       expect(visLayerError).toStrictEqual(expectedVisLayerError);
     });
     test('get no permission ErrorVisLayer', async () => {
@@ -135,7 +134,6 @@ describe('overlay_anomalies spec', () => {
         message: error.message,
       };
       const visLayerError = await getVisLayerError(error);
-      console.log('recievedError: ' + JSON.stringify(visLayerError));
       expect(visLayerError).toStrictEqual(expectedVisLayerError);
     });
     test('get fetch issue ErrorVisLayer', async () => {
@@ -145,7 +143,6 @@ describe('overlay_anomalies spec', () => {
         message: error.message,
       };
       const visLayerError = await getVisLayerError(error);
-      console.log('recievedError: ' + JSON.stringify(visLayerError));
       expect(visLayerError).toStrictEqual(expectedVisLayerError);
     });
   });
