+++ public/expressions/overlay_anomalies.ts
@@ -3,46 +3,38 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import { get, isEmpty } from 'lodash';
+import { get } from 'lodash';
 import { i18n } from '@osd/i18n';
 import { ExpressionFunctionDefinition } from '../../../../src/plugins/expressions/public';
 import {
   VisLayerTypes,
   VisLayers,
   ExprVisLayers,
-  PluginResource,
 } from '../../../../src/plugins/vis_augmenter/public';
 import {
   TimeRange,
   calculateBounds,
 } from '../../../../src/plugins/data/common';
-import {
-  getAnomalySummaryQuery,
-  parsePureAnomalies,
-} from '../pages/utils/anomalyResultUtils';
-import { AD_NODE_API } from '../../utils/constants';
-import { AnomalyData } from '../models/interfaces';
-import { getClient } from '../services';
-import {
-  PointInTimeEventsVisLayer,
-  VisLayerError,
-  VisLayerErrorTypes,
-} from '../../../../src/plugins/vis_augmenter/public';
+import { PointInTimeEventsVisLayer } from '../../../../src/plugins/vis_augmenter/public';
 import { PLUGIN_NAME } from '../utils/constants';
 import {
   CANT_FIND_KEY_WORD,
   DOES_NOT_HAVE_PERMISSIONS_KEY_WORD,
-  NO_PERMISSIONS_KEY_WORD,
 } from '../../server/utils/helpers';
 import {
   DETECTOR_HAS_BEEN_DELETED,
-  ORIGIN_PLUGIN_VIS_LAYER,
   OVERLAY_ANOMALIES,
   PLUGIN_EVENT_TYPE,
   START_OR_END_TIME_INVALID_ERROR,
   TYPE_OF_EXPR_VIS_LAYERS,
   VIS_LAYER_PLUGIN_TYPE,
 } from './constants';
+import {
+  convertAnomaliesToPointInTimeEventsVisLayer,
+  getAnomalies,
+  getDetectorResponse,
+  getVisLayerError,
+} from './helpers';
 
 type Input = ExprVisLayers;
 type Output = Promise<ExprVisLayers>;
@@ -54,104 +46,6 @@ interface Arguments {
 
 export type OverlayAnomaliesExpressionFunctionDefinition =
   ExpressionFunctionDefinition<Name, Input, Arguments, Output>;
-
-// This gets all the needed anomalies for the given detector ID and time range
-export const getAnomalies = async (
-  detectorId: string,
-  startTime: number,
-  endTime: number
-): Promise<AnomalyData[]> => {
-  const anomalySummaryQuery = getAnomalySummaryQuery(
-    startTime,
-    endTime,
-    detectorId,
-    undefined,
-    true
-  );
-
-  const anomalySummaryResponse = await getClient().post(
-    `..${AD_NODE_API.DETECTOR}/results/_search`,
-    {
-      body: JSON.stringify(anomalySummaryQuery),
-    }
-  );
-
-  return parsePureAnomalies(anomalySummaryResponse);
-};
-
-export const getDetectorResponse = async (detectorId: string) => {
-  const resp = await getClient().get(`..${AD_NODE_API.DETECTOR}/${detectorId}`);
-  return resp;
-};
-
-// This takes anomalies and returns them as vis layer of type PointInTimeEvents
-export const convertAnomaliesToPointInTimeEventsVisLayer = (
-  anomalies: AnomalyData[],
-  ADPluginResource: PluginResource
-): PointInTimeEventsVisLayer => {
-  const events = anomalies.map((anomaly: AnomalyData) => {
-    return {
-      timestamp: anomaly.startTime + (anomaly.endTime - anomaly.startTime) / 2,
-      metadata: {},
-    };
-  });
-  return {
-    originPlugin: ORIGIN_PLUGIN_VIS_LAYER,
-    type: VisLayerTypes.PointInTimeEvents,
-    pluginResource: ADPluginResource,
-    events: events,
-  } as PointInTimeEventsVisLayer;
-};
-
-const checkIfPermissionErrors = (error): boolean => {
-  return typeof error === 'string'
-    ? error.includes(NO_PERMISSIONS_KEY_WORD) ||
-        error.includes(DOES_NOT_HAVE_PERMISSIONS_KEY_WORD)
-    : get(error, 'message', '').includes(NO_PERMISSIONS_KEY_WORD) ||
-        get(error, 'message', '').includes(DOES_NOT_HAVE_PERMISSIONS_KEY_WORD);
-};
-
-//Helps convert any possible errors into either permission, deletion or fetch related failures
-export const getVisLayerError = (error): VisLayerError => {
-  let visLayerError: VisLayerError = {} as VisLayerError;
-  if (checkIfPermissionErrors(error)) {
-    visLayerError = {
-      type: VisLayerErrorTypes.PERMISSIONS_FAILURE,
-      message:
-        error === 'string'
-          ? error
-          : error instanceof Error
-          ? get(error, 'message', '')
-          : '',
-    };
-  } else if (
-    typeof error === 'string'
-      ? error.includes(DETECTOR_HAS_BEEN_DELETED)
-      : get(error, 'message', '').includes(DETECTOR_HAS_BEEN_DELETED)
-  ) {
-    console.log('inside resource deleted');
-    visLayerError = {
-      type: VisLayerErrorTypes.RESOURCE_DELETED,
-      message:
-        error === 'string'
-          ? error
-          : error instanceof Error
-          ? get(error, 'message', '')
-          : '',
-    };
-  } else {
-    visLayerError = {
-      type: VisLayerErrorTypes.FETCH_FAILURE,
-      message:
-        error === 'string'
-          ? error
-          : error instanceof Error
-          ? get(error, 'message', '')
-          : '',
-    };
-  }
-  return visLayerError;
-};
 
 /*
  * This function defines the Anomaly Detection expression function of type vis_layers.
